rename all js modules to mjs
