+++ root/static/js/main.mjs
@@ -1,10 +1,10 @@
-import './autocomplete.js';
-import './tablesorter.js';
+import './autocomplete.mjs';
+import './tablesorter.mjs';
 import './cpan.js';
 import './github.js';
 import './dropdown.js';
 import './profile.js';
-import './recaptcha.js';
+import './recaptcha.mjs';
 import './search.js';
 import 'bootstrap/js/dropdown.js';
 import 'bootstrap/js/collapse.js';
