+++ root/static/js/recaptcha.mjs
@@ -1,5 +1,3 @@
-'use strict';
-
 function recaptcha_prepare() {
     return new Promise((resolve, reject) => {
         const recaptcha_script = document.createElement('script');
